refactor: move Strictness and has_fatal_errors to preflight/strictness.py
Breaks the preflight -> run dependency cycle. These are preflight
error interpretation policies, not run-specific logic.
