+++ nbkp/run/pipeline.py
@@ -8,7 +8,6 @@ caller's responsibility.
 
 from __future__ import annotations
 
-import enum
 from dataclasses import dataclass
 from typing import Any, Callable
 
@@ -20,6 +19,7 @@ from ..preflight import (
     VolumeStatus,
     check_all_syncs,
 )
+from ..preflight.strictness import Strictness, has_fatal_errors
 from ..sync.rsync import ProgressMode
 from ..sync.runner import SyncResult, run_all_syncs
 
@@ -57,44 +57,6 @@ def is_expected_skip(
     """
     ss = sync_statuses.get(result.sync_slug)
     return ss is not None and ss.is_expected_inactive()
-
-
-class Strictness(str, enum.Enum):
-    """Controls how preflight errors affect the exit code.
-
-    - ``IGNORE_NONE``: All errors are fatal — any inactive sync
-      (including missing sentinels) aborts the run.
-    - ``IGNORE_INACTIVE``: Expected-inactive errors (missing sentinels,
-      unreachable hosts) are silently skipped; infrastructure errors
-      are still fatal.  This is the default.
-    - ``IGNORE_ALL``: All preflight errors are ignored — only sync
-      execution failures cause a non-zero exit.
-    """
-
-    IGNORE_NONE = "ignore-none"
-    IGNORE_INACTIVE = "ignore-inactive"
-    IGNORE_ALL = "ignore-all"
-
-
-def has_fatal_errors(
-    sync_statuses: dict[str, SyncStatus],
-    *,
-    strictness: Strictness = Strictness.IGNORE_INACTIVE,
-) -> bool:
-    """Return True if any sync has errors that should abort the run.
-
-    See :class:`Strictness` for the three modes.
-    """
-    match strictness:
-        case Strictness.IGNORE_NONE:
-            return any(not s.active for s in sync_statuses.values())
-        case Strictness.IGNORE_INACTIVE:
-            return any(
-                not s.active and not s.is_expected_inactive()
-                for s in sync_statuses.values()
-            )
-        case Strictness.IGNORE_ALL:
-            return False
 
 
 def check_and_run(
